backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.1.0
django-filter==25.1
drf-spectacular==0.27.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.28.2
